+++ src/olmo_core/data/data_loader.py
@@ -208,6 +208,14 @@ class DataLoaderBase(ABC):
         self.batches_processed = 0
         self.tokens_processed = 0
 
+    @abstractmethod
+    def get_mock_batch(self) -> Dict[str, Any]:
+        """
+        Return a batch with arbitrary data. This can just be random data as it's only used by the
+        trainer to do a dry-run of the forward and backward pass before training officially starts.
+        """
+        raise NotImplementedError
+
 
 class NumpyDataLoaderBase(DataLoaderBase):
     """
@@ -420,6 +428,13 @@ class NumpyDataLoaderBase(DataLoaderBase):
             raise ValueError(f"'epoch' must be at least 1, got {epoch}")
         self._epoch = epoch
         self.build_and_save_global_indices(in_memory=in_memory)
+
+    def get_mock_batch(self) -> Dict[str, Any]:
+        num_instances = self.rank_batch_size // self.dataset.max_sequence_length
+        input_ids = torch.randint(
+            0, self.dataset.vocab_size, (num_instances, self.dataset.max_sequence_length)
+        )
+        return {"input_ids": input_ids}
 
     def _iter_batches(self) -> Iterable[Dict[str, Any]]:
         return torch.utils.data.DataLoader(